requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.7
django-cloudinary-storage==0.3.0
django-elasticsearch-dsl==8.0
django-nested-admin==4.1.6
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
elastic-transport==8.13.0
elasticsearch==8.13.0
elasticsearch-dsl==8.13.0
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.11
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-monkey-business==1.1.0
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.11.0